refactor: move tasks and ui from main.py to seperate files
